dev_requirements.txt
-r requirements.txt

django-debug-toolbar==3.2
django-extensions==3.1.1